requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
langgraph>=0.2.60
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
python-dotenv>=1.0.1
